+++ src/views/tableView.ts
@@ -351,6 +351,7 @@ export class TableView {
                 <option value="">📥&nbsp;Save</option>
                 <option value="csv">,,, csv</option>
                 <option value="json">{} &nbsp;json</option>
+                <option value="html">&lt;&gt; &nbsp;html</option>
               </select>
               <vscode-button id="reload-button"
                 appearance="icon" aria-label="Reload Data">